using structs
