+++ run/main.go
@@ -9,6 +9,7 @@ import (
 	"github.com/paketo-buildpacks/packit/v2/scribe"
 
 	ubinodejsextension "github.com/paketo-community/ubi-nodejs-extension"
+	"github.com/paketo-community/ubi-nodejs-extension/internal/utils"
 )
 
 const IMAGES_JSON_PATH = "/etc/buildpacks/images.json"
@@ -16,7 +17,7 @@ const IMAGES_JSON_PATH = "/etc/buildpacks/images.json"
 func main() {
 	dependencyManager := postal.NewService(cargo.NewTransport())
 	logEmitter := scribe.NewEmitter(os.Stdout).WithLevel(os.Getenv("BP_LOG_LEVEL"))
-	duringBuildPermissions := ubinodejsextension.GetDuringBuildPermissions("/etc/passwd")
+	duringBuildPermissions := utils.GetDuringBuildPermissions("/etc/passwd")
 
 	packit.RunExtension(
 		ubinodejsextension.Detect(),
